+++ mlir/include/mlir/Dialect/NVVM/IR/CMakeLists.txt
@@ -1,0 +1,21 @@
+set(LLVM_TARGET_DEFINITIONS BasicPtxBuilderInterface.td)
+mlir_tablegen(BasicPtxBuilderInterface.h.inc -gen-op-interface-decls)
+mlir_tablegen(BasicPtxBuilderInterface.cpp.inc -gen-op-interface-defs)
+add_mlir_dialect_tablegen_target(MLIRBasicPtxBuilderInterfaceIncGen)
+
+set(LLVM_TARGET_DEFINITIONS NVVMRequiresSMTraits.td)
+mlir_tablegen(NVVMRequiresSMTraits.h.inc -gen-op-interface-decls)
+mlir_tablegen(NVVMRequiresSMTraits.cpp.inc -gen-op-interface-defs)
+add_mlir_dialect_tablegen_target(MLIRNVVMRequiresSMTraitsIncGen)
+
+add_mlir_dialect(NVVMOps nvvm)
+add_mlir_doc(NVVMOps NVVMOps Dialects/ -gen-op-doc -dialect=nvvm)
+set(LLVM_TARGET_DEFINITIONS NVVMOps.td)
+mlir_tablegen(NVVMConversions.inc -gen-llvmir-conversions)
+mlir_tablegen(NVVMFromLLVMIRConversions.inc -gen-intr-from-llvmir-conversions)
+mlir_tablegen(NVVMConvertibleLLVMIRIntrinsics.inc -gen-convertible-llvmir-intrinsics)
+mlir_tablegen(NVVMOpsEnums.h.inc -gen-enum-decls)
+mlir_tablegen(NVVMOpsEnums.cpp.inc -gen-enum-defs)
+mlir_tablegen(NVVMOpsAttributes.h.inc -gen-attrdef-decls -attrdefs-dialect=nvvm)
+mlir_tablegen(NVVMOpsAttributes.cpp.inc -gen-attrdef-defs -attrdefs-dialect=nvvm)
+add_mlir_dialect_tablegen_target(MLIRNVVMConversionsIncGen)
