+++ mlir/lib/Conversion/GPUToNVVM/WmmaOpsToNvvm.cpp
@@ -15,7 +15,7 @@
 #include "mlir/Conversion/LLVMCommon/Pattern.h"
 #include "mlir/Dialect/GPU/IR/GPUDialect.h"
 #include "mlir/Dialect/LLVMIR/LLVMDialect.h"
-#include "mlir/Dialect/LLVMIR/NVVMDialect.h"
+#include "mlir/Dialect/NVVM/IR/NVVMDialect.h"
 #include "mlir/IR/TypeUtilities.h"
 #include "mlir/IR/Types.h"
