+++ mlir/lib/Dialect/NVVM/CMakeLists.txt
@@ -1,1 +1,2 @@
+add_subdirectory(IR)
 add_subdirectory(Transforms)
