+++ mlir/lib/Dialect/NVVM/Transforms/OptimizeForNVVM.cpp
@@ -8,7 +8,7 @@
 
 #include "mlir/Dialect/NVVM/Transforms/OptimizeForNVVM.h"
 
-#include "mlir/Dialect/LLVMIR/NVVMDialect.h"
+#include "mlir/Dialect/NVVM/IR/NVVMDialect.h"
 #include "mlir/IR/Builders.h"
 #include "mlir/IR/PatternMatch.h"
 #include "mlir/Pass/Pass.h"
