+++ utils/bazel/llvm-project-overlay/mlir/BUILD.bazel
@@ -5468,10 +5468,7 @@ cc_library(
             "lib/Dialect/LLVMIR/IR/*AMX*.h",
             "lib/Dialect/LLVMIR/IR/*ArmSVE*.cpp",
             "lib/Dialect/LLVMIR/IR/*ArmSVE*.h",
-            "lib/Dialect/LLVMIR/IR/NVVM*.cpp",
-            "lib/Dialect/LLVMIR/IR/NVVM*.h",
             "lib/Dialect/LLVMIR/IR/ROCDL*.cpp",
-            "lib/Dialect/LLVMIR/IR/BasicPtxBuilderInterface.cpp",
             "lib/Dialect/LLVMIR/IR/ROCDL*.h",
             "lib/Dialect/LLVMIR/IR/VCIX*.cpp",
             "lib/Dialect/LLVMIR/IR/VCIX*.h",
@@ -5486,8 +5483,6 @@ cc_library(
         exclude = [
             "include/mlir/Dialect/LLVMIR/*AMX*.h",
             "include/mlir/Dialect/LLVMIR/*ArmSVE*.h",
-            "include/mlir/Dialect/LLVMIR/NVVM*.h",
-            "include/mlir/Dialect/LLVMIR/BasicPtxBuilderInterface.h",
             "include/mlir/Dialect/LLVMIR/ROCDL*.h",
             "include/mlir/Dialect/LLVMIR/VCIX*.h",
             "include/mlir/Dialect/LLVMIR/*X86*.h",
@@ -6393,14 +6388,14 @@ gentbl_cc_library(
 cc_library(
     name = "NVVMDialect",
     srcs = [
-        "lib/Dialect/LLVMIR/IR/BasicPtxBuilderInterface.cpp",
-        "lib/Dialect/LLVMIR/IR/NVVMDialect.cpp",
-        "lib/Dialect/LLVMIR/IR/NVVMRequiresSMTraits.cpp",
+        "lib/Dialect/NVVM/IR/BasicPtxBuilderInterface.cpp",
+        "lib/Dialect/NVVM/IR/NVVMDialect.cpp",
+        "lib/Dialect/NVVM/IR/NVVMRequiresSMTraits.cpp",
     ],
     hdrs = [
-        "include/mlir/Dialect/LLVMIR/BasicPtxBuilderInterface.h",
-        "include/mlir/Dialect/LLVMIR/NVVMDialect.h",
-        "include/mlir/Dialect/LLVMIR/NVVMRequiresSMTraits.h",
+        "include/mlir/Dialect/NVVM/IR/BasicPtxBuilderInterface.h",
+        "include/mlir/Dialect/NVVM/IR/NVVMDialect.h",
+        "include/mlir/Dialect/NVVM/IR/NVVMRequiresSMTraits.h",
     ],
     includes = ["include"],
     deps = [
@@ -6497,7 +6492,7 @@ cc_library(
 
 td_library(
     name = "BasicPtxBuilderIntTdFiles",
-    srcs = ["include/mlir/Dialect/LLVMIR/BasicPtxBuilderInterface.td"],
+    srcs = ["include/mlir/Dialect/NVVM/IR/BasicPtxBuilderInterface.td"],
     includes = ["include"],
     deps = [
         ":GPUOpsTdFiles",
@@ -6508,7 +6503,7 @@ td_library(
 
 td_library(
     name = "NVVMRequiresSMTraitsTdFiles",
-    srcs = ["include/mlir/Dialect/LLVMIR/NVVMRequiresSMTraits.td"],
+    srcs = ["include/mlir/Dialect/NVVM/IR/NVVMRequiresSMTraits.td"],
     includes = ["include"],
     deps = [
         ":LLVMOpsTdFiles",
@@ -6519,9 +6514,9 @@ td_library(
 td_library(
     name = "NVVMOpsTdFiles",
     srcs = [
-        "include/mlir/Dialect/LLVMIR/NVVMDialect.td",
-        "include/mlir/Dialect/LLVMIR/NVVMEnums.td",
-        "include/mlir/Dialect/LLVMIR/NVVMOps.td",
+        "include/mlir/Dialect/NVVM/IR/NVVMDialect.td",
+        "include/mlir/Dialect/NVVM/IR/NVVMEnums.td",
+        "include/mlir/Dialect/NVVM/IR/NVVMOps.td",
     ],
     includes = ["include"],
     deps = [
@@ -6537,11 +6532,11 @@ td_library(
 gentbl_cc_library(
     name = "BasicPtxBuilderIntGen",
     tbl_outs = {
-        "include/mlir/Dialect/LLVMIR/BasicPtxBuilderInterface.h.inc": ["-gen-op-interface-decls"],
-        "include/mlir/Dialect/LLVMIR/BasicPtxBuilderInterface.cpp.inc": ["-gen-op-interface-defs"],
+        "include/mlir/Dialect/NVVM/IR/BasicPtxBuilderInterface.h.inc": ["-gen-op-interface-decls"],
+        "include/mlir/Dialect/NVVM/IR/BasicPtxBuilderInterface.cpp.inc": ["-gen-op-interface-defs"],
     },
     tblgen = ":mlir-tblgen",
-    td_file = "include/mlir/Dialect/LLVMIR/BasicPtxBuilderInterface.td",
+    td_file = "include/mlir/Dialect/NVVM/IR/BasicPtxBuilderInterface.td",
     deps = [
         ":BasicPtxBuilderIntTdFiles",
         ":GPUOpsTdFiles",
@@ -6552,64 +6547,64 @@ gentbl_cc_library(
 gentbl_cc_library(
     name = "NVVMRequiresSMTraitsIncGen",
     tbl_outs = {
-        "include/mlir/Dialect/LLVMIR/NVVMRequiresSMTraits.h.inc": ["-gen-op-interface-decls"],
-        "include/mlir/Dialect/LLVMIR/NVVMRequiresSMTraits.cpp.inc": ["-gen-op-interface-defs"],
+        "include/mlir/Dialect/NVVM/IR/NVVMRequiresSMTraits.h.inc": ["-gen-op-interface-decls"],
+        "include/mlir/Dialect/NVVM/IR/NVVMRequiresSMTraits.cpp.inc": ["-gen-op-interface-defs"],
     },
     tblgen = ":mlir-tblgen",
-    td_file = "include/mlir/Dialect/LLVMIR/NVVMRequiresSMTraits.td",
+    td_file = "include/mlir/Dialect/NVVM/IR/NVVMRequiresSMTraits.td",
     deps = [":NVVMRequiresSMTraitsTdFiles"],
 )
 
 gentbl_cc_library(
     name = "NVVMOpsIncGen",
     tbl_outs = {
-        "include/mlir/Dialect/LLVMIR/NVVMOps.h.inc": ["-gen-op-decls"],
-        "include/mlir/Dialect/LLVMIR/NVVMOps.cpp.inc": ["-gen-op-defs"],
-        "include/mlir/Dialect/LLVMIR/NVVMOpsDialect.h.inc": [
+        "include/mlir/Dialect/NVVM/IR/NVVMOps.h.inc": ["-gen-op-decls"],
+        "include/mlir/Dialect/NVVM/IR/NVVMOps.cpp.inc": ["-gen-op-defs"],
+        "include/mlir/Dialect/NVVM/IR/NVVMOpsDialect.h.inc": [
             "-gen-dialect-decls",
             "-dialect=nvvm",
         ],
-        "include/mlir/Dialect/LLVMIR/NVVMOpsDialect.cpp.inc": [
+        "include/mlir/Dialect/NVVM/IR/NVVMOpsDialect.cpp.inc": [
             "-gen-dialect-defs",
             "-dialect=nvvm",
         ],
-        "include/mlir/Dialect/LLVMIR/NVVMOpsEnums.h.inc": ["-gen-enum-decls"],
-        "include/mlir/Dialect/LLVMIR/NVVMOpsEnums.cpp.inc": ["-gen-enum-defs"],
-        "include/mlir/Dialect/LLVMIR/NVVMOpsAttributes.h.inc": [
+        "include/mlir/Dialect/NVVM/IR/NVVMOpsEnums.h.inc": ["-gen-enum-decls"],
+        "include/mlir/Dialect/NVVM/IR/NVVMOpsEnums.cpp.inc": ["-gen-enum-defs"],
+        "include/mlir/Dialect/NVVM/IR/NVVMOpsAttributes.h.inc": [
             "-gen-attrdef-decls",
             "-attrdefs-dialect=nvvm",
         ],
-        "include/mlir/Dialect/LLVMIR/NVVMOpsAttributes.cpp.inc": [
+        "include/mlir/Dialect/NVVM/IR/NVVMOpsAttributes.cpp.inc": [
             "-gen-attrdef-defs",
             "-attrdefs-dialect=nvvm",
         ],
     },
     tblgen = ":mlir-tblgen",
-    td_file = "include/mlir/Dialect/LLVMIR/NVVMOps.td",
+    td_file = "include/mlir/Dialect/NVVM/IR/NVVMOps.td",
     deps = [":NVVMOpsTdFiles"],
 )
 
 gentbl_cc_library(
     name = "NVVMConversionIncGen",
-    tbl_outs = {"include/mlir/Dialect/LLVMIR/NVVMConversions.inc": ["-gen-llvmir-conversions"]},
+    tbl_outs = {"include/mlir/Dialect/NVVM/IR/NVVMConversions.inc": ["-gen-llvmir-conversions"]},
     tblgen = ":mlir-tblgen",
-    td_file = "include/mlir/Dialect/LLVMIR/NVVMOps.td",
+    td_file = "include/mlir/Dialect/NVVM/IR/NVVMOps.td",
     deps = [":NVVMOpsTdFiles"],
 )
 
 gentbl_cc_library(
     name = "NVVMFromLLVMIRConversionsIncGen",
-    tbl_outs = {"include/mlir/Dialect/LLVMIR/NVVMFromLLVMIRConversions.inc": ["-gen-intr-from-llvmir-conversions"]},
+    tbl_outs = {"include/mlir/Dialect/NVVM/IR/NVVMFromLLVMIRConversions.inc": ["-gen-intr-from-llvmir-conversions"]},
     tblgen = ":mlir-tblgen",
-    td_file = "include/mlir/Dialect/LLVMIR/NVVMOps.td",
+    td_file = "include/mlir/Dialect/NVVM/IR/NVVMOps.td",
     deps = [":NVVMOpsTdFiles"],
 )
 
 gentbl_cc_library(
     name = "NVVMConvertibleLLVMIRIntrinsicsIncGen",
-    tbl_outs = {"include/mlir/Dialect/LLVMIR/NVVMConvertibleLLVMIRIntrinsics.inc": ["-gen-convertible-llvmir-intrinsics"]},
+    tbl_outs = {"include/mlir/Dialect/NVVM/IR/NVVMConvertibleLLVMIRIntrinsics.inc": ["-gen-convertible-llvmir-intrinsics"]},
     tblgen = ":mlir-tblgen",
-    td_file = "include/mlir/Dialect/LLVMIR/NVVMOps.td",
+    td_file = "include/mlir/Dialect/NVVM/IR/NVVMOps.td",
     deps = [":NVVMOpsTdFiles"],
 )
